chore: move the function ascii_as_felt to starknet core
