+++ crates/starknet_api/src/transaction_hash_test.rs
@@ -3,25 +3,11 @@ use serde::{Deserialize, Serialize};
 use sha3::{Digest, Keccak256};
 use starknet_types_core::felt::Felt;
 
-use super::{
-    ascii_as_felt,
-    get_transaction_hash,
-    validate_transaction_hash,
-    CONSTRUCTOR_ENTRY_POINT_SELECTOR,
-};
+use super::{get_transaction_hash, validate_transaction_hash, CONSTRUCTOR_ENTRY_POINT_SELECTOR};
 use crate::block::BlockNumber;
 use crate::core::ChainId;
 use crate::test_utils::read_json_file;
 use crate::transaction::{Transaction, TransactionHash, TransactionOptions};
-
-#[test]
-fn test_ascii_as_felt() {
-    let sn_main_id = ChainId::Mainnet;
-    let sn_main_felt = ascii_as_felt(sn_main_id.to_string().as_str()).unwrap();
-    // This is the result of the Python snippet from the Chain-Id documentation.
-    let expected_sn_main = Felt::from(23448594291968334_u128);
-    assert_eq!(sn_main_felt, expected_sn_main);
-}
 
 #[test]
 fn test_constructor_selector() {
